MPEG TS reader: update TextST CodecPrivate format to what current MakeMKV uses
At the end of 2014 Mike from MakeMKV and me where debating how to store
TextST in Matroska. MakeMKV was already able to do that and used this as
the format for CodecPrivate:

• One byte: character set used according to "Blu-ray Disc Read-Only
  Format: table 5-22: Character code"
• n bytes: the "dialog style segment"
• 2 bytes: number of frames in the file

This information partially superfluous, partially harmful. We decided to
only allow UTF-8 in Matroska (we only had Blu-rays that used UTF-8
anyway; this gets rid of the first byte). We also decided that having a
frame counter served no purpose (the format doesn't need it, nor should
players require such information up front), and it made splitting &
joining operations harder than they have to be.

Therefore current versions of MakeMKV only include the "dialog style
segment" in CodecPrivate, and that's what mkvmerge will do, too.

The Matroska reader was already producing the new format.

+++ src/input/r_matroska.cpp
@@ -661,7 +661,7 @@ kax_reader_c::verify_hdmv_textst_subtitle_track(kax_track_t *t) {
   auto style_segment_start = old_style ? 1 : 0;
   auto style_segment_size  = get_uint16_be(&buf[style_segment_start + 1]);
 
-  if (t->private_size < (3 + style_segment_size + (old_style ? 1 + 2 : 0))) {
+  if (t->private_size < static_cast<unsigned int>(3 + style_segment_size + (old_style ? 1 + 2 : 0))) {
     mxwarn(boost::format(Y("matroska_reader: The CodecID for track %1% is '%2%', but the private codec data does not contain valid headers.\n")) % t->codec_id);
     return false;
   }
